feat(fw-esp32v3): level-4 RMT refill vector + head-to-head stress harness
The experiment proper. hli_refill (additive to the server build) swaps the
endpoint layer's refill path onto a hand-written Xtensa level-4 vector:

- vector.rs: __naked_level_4_interrupt via xtensa-lx-rt's PROVIDE seam
  (the sanctioned esp-hal-world hook), RMT routed to CPU interrupt 24
  (extern, level-triggered, level 4) through the peripheral's own matrix
  register. Call0 discipline, no windowed instructions, a2..a13+SAR saved
  to a static DRAM area, code+literals in IRAM, ack-all storm guard,
  rfi 4. Field offsets arrive as global_asm const operands from
  offset_of! on the lp-ws281x-hli contract structs — nothing transcribed.
- app.rs: the shared_driver-shaped thread side (configure/send/abort/
  telemetry); the start path runs the host-tested reference model, so the
  asm's spec executes on silicon every frame beside the asm itself.
  Telemetry prints the P4 line format + src=hli4.
- stress.rs (hli_stress, entrypoint-replacing like radio_ram_probe):
  radio-linked cells on one boot — L3 idle/scan then a live matrix switch
  to the level-4 vector, idle/scan again; 4x30 LEDs at the P4 baseline
  pace, C6's S2 scan generator.

Coexistence facts recorded in the module docs: esp-sync critical sections
are rsil 5, so level 4 escapes only the rsil<=3 class (embassy locks,
level<=3 handler execution) — which class dominates is what the harness
measures.

Feature off: section sizes byte-identical (verified against a stashed
build). Feature on vs telemetry build: .bss +816 B (bank + save area),
.data -1088 B / .rwtext -452 B (idle L3 refill path GC'd), flash +264 B.
Clippy matrix extended and green for all six configurations.

+++ lp-fw/lp-ws281x-hli/src/lib.rs
@@ -220,9 +220,9 @@ impl Default for HliChannel {
 pub struct HliBank {
     /// Address of the RMT `INT_ST` register (`raw & ena` — causes this
     /// firmware never enabled cannot appear).
-    pub int_st_addr: AtomicU32,
+    pub int_st_addr: AtomicUsize,
     /// Address of the RMT `INT_CLR` register (write-1-to-clear).
-    pub int_clr_addr: AtomicU32,
+    pub int_clr_addr: AtomicUsize,
     /// Union of every entry's three cause masks: the causes this handler owns.
     /// Everything pending under this mask is acknowledged in one write at
     /// entry, whether or not the channel is active — storm safety for a
@@ -240,8 +240,8 @@ impl HliBank {
     /// fill it in.
     pub const fn new() -> Self {
         Self {
-            int_st_addr: AtomicU32::new(0),
-            int_clr_addr: AtomicU32::new(0),
+            int_st_addr: AtomicUsize::new(0),
+            int_clr_addr: AtomicUsize::new(0),
             all_mask: AtomicU32::new(0),
             isr_entries: AtomicU32::new(0),
             channels: [const { HliChannel::new() }; HLI_CHANNELS],
